Fix Travis warning: multi-character character constant [-Wmultichar]
(based on cmangos/mangos-wotlk@478752e31)

+++ contrib/extractor/loadlib/adt.cpp
@@ -74,7 +74,7 @@ bool ADT_file::prepareLoadedData()
 
 bool adt_MHDR::prepareLoadedData()
 {
-    if (fcc != 'MHDR')
+    if (fcc != uint32('MHDR'))
         return false;
 
     if (size != sizeof(adt_MHDR) - 8)
@@ -93,7 +93,7 @@ bool adt_MHDR::prepareLoadedData()
 
 bool adt_MCIN::prepareLoadedData()
 {
-    if (fcc != 'MCIN')
+    if (fcc != uint32('MCIN'))
         return false;
 
     // Check cells data
@@ -107,7 +107,7 @@ bool adt_MCIN::prepareLoadedData()
 
 bool adt_MH2O::prepareLoadedData()
 {
-    if (fcc != 'MH2O')
+    if (fcc != uint32('MH2O'))
         return false;
 
     // Check liquid data
@@ -119,7 +119,7 @@ bool adt_MH2O::prepareLoadedData()
 
 bool adt_MCNK::prepareLoadedData()
 {
-    if (fcc != 'MCNK')
+    if (fcc != uint32('MCNK'))
         return false;
 
     // Check height map
@@ -134,7 +134,7 @@ bool adt_MCNK::prepareLoadedData()
 
 bool adt_MCVT::prepareLoadedData()
 {
-    if (fcc != 'MCVT')
+    if (fcc != uint32('MCVT'))
         return false;
 
     if (size != sizeof(adt_MCVT) - 8)
@@ -145,7 +145,7 @@ bool adt_MCVT::prepareLoadedData()
 
 bool adt_MCLQ::prepareLoadedData()
 {
-    if (fcc != 'MCLQ')
+    if (fcc != uint32('MCLQ'))
         return false;
 
     return true;
